Misc changes (#1710)
* Minor code refactoring for re-use.

* Add Class.FindVariable helper method.

* Upgrade to .NET 6.

(cherry picked from commit db7949b263e90158c0e8cb388bf243b0f80de4cb)
