+++ src/AST/Class.cs
@@ -291,6 +291,11 @@ namespace CppSharp.AST
                 .FirstOrDefault(m => m.USR == usr);
         }
 
+        public Variable FindVariable(string name)
+        {
+            return Variables.FirstOrDefault(m => m.Name == name);
+        }
+
         public override T Visit<T>(IDeclVisitor<T> visitor)
         {
             return visitor.VisitClassDecl(this);
